11.0 - openfield new ideas
* OF is now playable through a HTML5 version linked in the latest
journal entry

? playtesting an unfinished and super unpolished build was hard, but I
realized I was less interested in the feedback, and was more looking
for an excuse to talk about the project to someone else. Verbalizing
the core concepts of the game gave me the perspective I was looking for
to generate new ideas.

? the game is still stuck in a high-level thinking place.  it is a real
challenge for me to boil down these ideas in concise gameplay.