14.0 - new (and hopefully last) prototype
* In this post I talk extensively about the past month that has been
creatively dry, and where I’m headed next.

? Took me a long time to write this long post, as I suffered from some
kind of writer’s block. Somehow, because I wasn’t making anything of
substance in the engine, I could not write about it. This is not a good
habit, I believe I should double down on the writing when I feel like I
don’t know where I am going. Or at least to reflect and ponder on how I
feel lost.

? Writing about my new idea (the apartment one) made me realize how
easy it is to describe it, which is usually an early good sign. I could
also see that the message behind has a very interesting potential

? Might not be able to integrate it for next week, but I’d love to add
a commentary (voice-over or subs) on illusion in games, using that as
some sort of narrative progression)

? I’d also like to talk about Clement Greenberg. I wrote an essay on
him last fall on illusion in painting vs illusion in games. He an
interesting character, so that could be fun. I am definitely settled on
finishing this idea, and see if it’s still worth pursuing.