15.0 - new DLP prototype analysis
* first iteration of DLP is done! Like I wanted, it’s a first version
that could be pimped in many ways, but it was made so that it would
stand on its own and offer the complete core experience

? I’ve had this idea for two years, but doing it and playing it opened
the door for much more potential than I expected. I think I might
publish it as it and use the feedback to think about where I could go
next with this, or how I could apply this concept elsewhere.